+++ tests/core/Statiq.Core.Tests/Shortcodes/Html/TableFixture.cs
@@ -77,7 +77,7 @@ l=m nop
             {
                 // Given
                 TestExecutionContext context = new TestExecutionContext();
-                context.AddTypeConversion<string, int>(x => int.Parse(x));
+                context.TypeConverter.AddTypeConversion<string, int>(x => int.Parse(x));
                 TestDocument document = new TestDocument();
                 string content = @"
 1 2 ""3 4""